+++ CHANGELOG.md
@@ -18,16 +18,25 @@ Please file a bug if you notice a violation of semantic versioning.
 
 ## [Unreleased]
 ### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [1.0.15] - 2025-08-29
+- TAG: [v1.0.15][1.0.15t]
+- COVERAGE: 98.17% -- 2259/2301 lines in 20 files
+- BRANCH COVERAGE: 81.00% -- 908/1121 branches in 20 files
+- 76.03% documented
+### Added
 - kettle-release: early validation of identical set of copyright years in README.md and CHANGELOG.md, adds current year if missing, aborts on mismatch
 - kettle-release: update KLOC in README.md
 - kettle-release: update Rakefile.example with version and date
 ### Changed
 - kettle-release: print package name and version released as final line
-### Deprecated
-### Removed
 ### Fixed
 - include gemfiles/modular/*gemfile.example with packaged gem
-### Security
 
 ## [1.0.14] - 2025-08-28
 - TAG: [v1.0.14][1.0.14t]
@@ -241,7 +250,7 @@ Please file a bug if you notice a violation of semantic versioning.
   - Selecting will run the selected workflow via `act`
   - This may move to its own gem in the future.
 
-[Unreleased]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.14...HEAD
+[Unreleased]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.15...HEAD
 [1.0.0]: https://github.com/kettle-rb/kettle-dev/compare/a427c302df09cfe4253a7c8d400333f9a4c1a208...v1.0.0
 [1.0.0t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.0
 [1.0.1]: https://gitlab.com/kettle-rb/kettle-dev/-/compare/v1.0.0...v1.0.1
@@ -272,3 +281,5 @@ Please file a bug if you notice a violation of semantic versioning.
 [1.0.13t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.13
 [1.0.14]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.13...v1.0.14
 [1.0.14t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.14
+[1.0.15]: https://github.com/kettle-rb/kettle-dev/compare/v1.0.14...v1.0.15
+[1.0.15t]: https://github.com/kettle-rb/kettle-dev/releases/tag/v1.0.15